web-app/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1
supabase==1.0.3
bcrypt==4.0.1
python-jose==3.3.0python 
